Extract common code from webpack config and run concurrently
